Remove StoreVolatile.[ch]pp
