+++ jlm/llvm/Makefile.sub
@@ -34,7 +34,6 @@ libllvm_SOURCES = \
     jlm/llvm/ir/operators/Phi.cpp \
     jlm/llvm/ir/operators/sext.cpp \
     jlm/llvm/ir/operators/Store.cpp \
-    jlm/llvm/ir/operators/StoreVolatile.cpp \
     jlm/llvm/ir/print.cpp \
     jlm/llvm/ir/RvsdgModule.cpp \
     jlm/llvm/ir/ssa.cpp \
@@ -116,7 +115,6 @@ libllvm_HEADERS = \
 	jlm/llvm/ir/operators/theta.hpp \
 	jlm/llvm/ir/operators/delta.hpp \
 	jlm/llvm/ir/operators/Store.hpp \
-	jlm/llvm/ir/operators/StoreVolatile.hpp \
 	jlm/llvm/ir/operators/alloca.hpp \
 	jlm/llvm/ir/operators/call.hpp \
 	jlm/llvm/ir/operators/sext.hpp \
@@ -163,7 +161,6 @@ libllvm_TESTS += \
 	tests/jlm/llvm/ir/operators/TestPhi \
 	tests/jlm/llvm/ir/operators/test-sext \
 	tests/jlm/llvm/ir/operators/StoreTests \
-	tests/jlm/llvm/ir/operators/StoreVolatileTests \
 	tests/jlm/llvm/ir/test-aggregation \
 	tests/jlm/llvm/ir/test-cfg \
 	tests/jlm/llvm/ir/test-cfg-node \
