+++ jlm/llvm/ir/operators.hpp
@@ -18,7 +18,6 @@
 #include <jlm/llvm/ir/operators/Phi.hpp>
 #include <jlm/llvm/ir/operators/sext.hpp>
 #include <jlm/llvm/ir/operators/Store.hpp>
-#include <jlm/llvm/ir/operators/StoreVolatile.hpp>
 #include <jlm/llvm/ir/operators/theta.hpp>
 
 #endif
