+++ tests/jlm/llvm/frontend/llvm/StoreTests.cpp
@@ -9,7 +9,6 @@
 #include <jlm/llvm/frontend/LlvmModuleConversion.hpp>
 #include <jlm/llvm/ir/operators/operators.hpp>
 #include <jlm/llvm/ir/operators/Store.hpp>
-#include <jlm/llvm/ir/operators/StoreVolatile.hpp>
 #include <jlm/llvm/ir/print.hpp>
 
 #include <llvm/IR/BasicBlock.h>
